Implement Associated arrays.
The biggest difference between how PICK and PHP handle arrays is in how they both handle recurring data.

If you take the simplest array in PICK, it is basically fields are rows, and columns are the related fields across the page.

Where as in PHP it is Columns are the fields, and the rows are the different values.

The new uObject::fetchAssoc allows you to link the fields together into a array similar to a PHP array which can be iterated as normal.